Upgrade android_splash_screen Gradle configuration
This commit modernizes the Gradle setup for the `android_splash_screen` sample.

Key changes include:
*   Upgrading Gradle wrapper from version 6.7 to 8.7.
*   Migrating to the plugins block in `settings.gradle` and `app/build.gradle` for applying Android, Kotlin, and Flutter Gradle plugins.
*   Removing the outdated `buildscript` block from the root `build.gradle`.
*   Updating `compileSdkVersion` and `targetSdkVersion` to 34.
*   Setting `minSdkVersion` to `flutter.minSdkVersion`.
*   Adding the `namespace` property as required by newer Android Gradle Plugin versions.
*   Refactoring Kotlin constant declarations in `MainActivity.kt`.

+++ android_splash_screen/android/app/src/main/kotlin/com/example/splash_screen_sample/MainActivity.kt
@@ -203,9 +203,9 @@ class MainActivity : FlutterActivity() {
   }
 
   private companion object {
-    const val SPLASHSCREEN_ALPHA_ANIMATION_DURATION = 500 as Long
-    const val SPLASHSCREEN_TY_ANIMATION_DURATION = 500 as Long
-    const val SPLASHSCREEN_FINAL_ANIMATION_ALPHA_ANIMATION_DURATION = 250 as Long
-    const val WAIT_FOR_AVD_TO_FINISH = false
+      const val SPLASHSCREEN_ALPHA_ANIMATION_DURATION = 500L
+      const val SPLASHSCREEN_TY_ANIMATION_DURATION = 500L
+      const val SPLASHSCREEN_FINAL_ANIMATION_ALPHA_ANIMATION_DURATION = 250L
+      const val WAIT_FOR_AVD_TO_FINISH = false
   }
 }
